xmss: mv KeyPair as a container
